requirements.txt
langchain
flask
flask-sqlalchemy
langchain
openai
pandas
python-dotenv
sqlalchemy
gunicorn
psycopg2-binary